refactor: migrate agent package
