+++ frontend/apps/app/components/Chat/Chat.tsx
@@ -1,7 +1,6 @@
 'use client'
 
-import type { TableGroupData } from '@/app/lib/schema/convertSchemaToText'
-import type { Schema } from '@liam-hq/db-structure'
+import type { Schema, TableGroup } from '@liam-hq/db-structure'
 import type { Tables } from '@liam-hq/db/supabase/database.types'
 import type { FC } from 'react'
 import { useCallback, useEffect, useRef, useState } from 'react'
@@ -28,7 +27,7 @@ import type { ChatEntry } from './types/chatTypes'
 
 interface Props {
   schemaData: Schema
-  tableGroups?: Record<string, TableGroupData>
+  tableGroups?: Record<string, TableGroup>
   designSessionId?: string
   organizationId?: string
   buildingSchemaId: string
