--- frontend/apps/app/libs/schema/index.ts
@@ -1,3 +1,0 @@
-export { operationsSchema } from './operationsSchema'
-export { createNewVersion } from './createNewVersion'
-export { applyPatchOperations } from './applyPatchOperations'
