+++ frontend/apps/app/package.json
@@ -12,11 +12,8 @@
     "@codemirror/state": "6.5.2",
     "@codemirror/view": "6.36.8",
     "@electric-sql/pglite": "0.3.2",
-    "@langchain/community": "0.3.43",
-    "@langchain/core": "0.3.57",
-    "@langchain/langgraph": "0.3.0",
-    "@langchain/openai": "0.5.11",
     "@lezer/highlight": "1.2.1",
+    "@liam-hq/agent": "workspace:*",
     "@liam-hq/db": "workspace:*",
     "@liam-hq/db-structure": "workspace:*",
     "@liam-hq/erd-core": "workspace:*",
@@ -36,9 +33,7 @@
     "codemirror": "6.0.1",
     "date-fns": "4.1.0",
     "dotenv": "16.5.0",
-    "fast-json-patch": "3.1.1",
     "flags": "4.0.1",
-    "langfuse-langchain": "3.37.2",
     "langfuse-vercel": "3.37.2",
     "lucide-react": "0.511.0",
     "next": "15.3.2",
