+++ frontend/internal-packages/agent/package.json
@@ -3,7 +3,19 @@
   "private": true,
   "version": "0.1.0",
   "main": "src/index.ts",
-  "dependencies": {},
+  "dependencies": {
+    "@langchain/community": "0.3.43",
+    "@langchain/core": "0.3.57",
+    "@langchain/langgraph": "0.3.0",
+    "@langchain/openai": "0.5.11",
+    "@liam-hq/db": "workspace:*",
+    "@liam-hq/db-structure": "workspace:*",
+    "fast-json-patch": "3.1.1",
+    "langfuse-langchain": "3.37.2",
+    "next": "15.3.2",
+    "ts-pattern": "5.7.0",
+    "valibot": "1.1.0"
+  },
   "devDependencies": {
     "@biomejs/biome": "1.9.4",
     "@liam-hq/configs": "workspace:*",
@@ -13,7 +25,6 @@
     "vitest": "3.1.4"
   },
   "scripts": {
-    "build": "tsc",
     "dev": "tsc --watch",
     "fmt": "concurrently \"pnpm:fmt:*\"",
     "fmt:biome": "biome check --write --unsafe .",
