--- frontend/internal-packages/agent/src/agent.ts
@@ -1,4 +1,0 @@
-// Agent implementation will go here
-export const agent = {
-  // TODO: Add agent implementation
-}
