+++ frontend/internal-packages/agent/src/chat/workflow/nodes/answerGenerationNode.ts
@@ -1,12 +1,12 @@
+import type { Operation } from 'fast-json-patch'
+import * as v from 'valibot'
 import {
   type AgentName,
   createPromptVariables,
   getAgent,
-} from '@/lib/langchain'
-import { createNewVersion } from '@/libs/schema/createNewVersion'
-import { operationsSchema } from '@/libs/schema/operationsSchema'
-import type { Operation } from 'fast-json-patch'
-import * as v from 'valibot'
+} from '../../../langchain'
+import { createNewVersion } from '../../../utils/createNewVersion'
+import { operationsSchema } from '../../../utils/operationsSchema'
 import type { WorkflowState } from '../types'
 
 interface PreparedAnswerGeneration {
