+++ frontend/internal-packages/agent/src/chat/workflow/nodes/validationNode.ts
@@ -1,4 +1,4 @@
-import { convertSchemaToText } from '@/app/lib/schema/convertSchemaToText'
+import { convertSchemaToText } from '../../../utils/convertSchemaToText'
 import type { WorkflowState } from '../types'
 
 export const validationNode = async (
