+++ frontend/internal-packages/agent/src/chat/workflow/services/stateManager.ts
