+++ frontend/internal-packages/agent/src/db/server.ts
@@ -1,0 +1,43 @@
+import { createServerClient } from '@liam-hq/db'
+import { cookies } from 'next/headers'
+
+// FIXME: This file was copied from @liam-hq/app.
+// It needs to be reviewed and eventually removed to eliminate the dependency on @liam-hq/db,
+// which is necessary to allow calls from the benchmarker.
+
+/**
+ * Create a Supabase client
+ * @param options Options
+ * @param options.useServiceRole Whether to use the service role key (true to bypass RLS)
+ * @returns Supabase client
+ */
+export async function createClient({
+  useServiceRole = false,
+}: {
+  useServiceRole?: boolean
+} = {}) {
+  const cookieStore = await cookies()
+
+  // Use the service role key if specified and available in environment variables
+  const apiKey =
+    useServiceRole && process.env.SUPABASE_SERVICE_ROLE_KEY
+      ? process.env.SUPABASE_SERVICE_ROLE_KEY
+      : (process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY ?? '')
+
+  return createServerClient(
+    process.env.NEXT_PUBLIC_SUPABASE_URL ?? '',
+    apiKey,
+    {
+      cookies: {
+        getAll() {
+          return cookieStore.getAll()
+        },
+        setAll(cookiesToSet) {
+          for (const { name, value, options } of cookiesToSet) {
+            cookieStore.set(name, value, options)
+          }
+        },
+      },
+    },
+  )
+}
