+++ frontend/internal-packages/agent/src/index.ts
@@ -1,1 +1,4 @@
-export * from './agent'
+export * from './chat/chatProcessor'
+export * from './utils/applyPatchOperations'
+export * from './utils/operationsSchema'
+export * from './utils/createNewVersion'
