+++ frontend/internal-packages/agent/src/langchain/agents/databaseSchemaBuildAgent/prompts.ts
