+++ frontend/internal-packages/agent/src/utils/__tests__/applyPatchOperations.test.ts
