+++ frontend/internal-packages/agent/src/utils/applyPatchOperations.ts
