+++ frontend/internal-packages/agent/src/utils/operationsSchema.ts
