+++ frontend/internal-packages/agent/src/vectorstore/index.ts
@@ -1,0 +1,2 @@
+export * from './supabaseVectorStore'
+export * from './syncSchemaVectorStore'
