+++ frontend/internal-packages/agent/src/vectorstore/syncSchemaVectorStore.ts
